Time: 34 ms (48.42%) | Memory: 15.7 MB (14.11%) - LeetSync

+++ 3195-separate-black-and-white-balls/separate-black-and-white-balls.cpp
@@ -1,14 +1,14 @@
 class Solution {
 public:
     long long minimumSteps(string s) {
-        long long count1=0;
-        for(long long i=0;i<s.length();i++){
+        int count1=0;
+        for(int i=0;i<s.length();i++){
             if(s[i]=='1'){
                 count1++;
             }
         }
         long long answer=0;
-        long long behind=0;
+        int behind=0;
         for(long long i=0;i<s.length();i++){
             if(s[i]=='1'){
                 long long a =s.length() -(count1-behind);
